Move jitbuilder linking to libwabt
Prior to this change, jitbuilder was being linked manually by every tool
that used libwabt. This had several issues, the most significant being
that it did not allow any libwabt source files to include jitbuilder
headers, something that will be extremely important.

A better solution is to add jitbuilder as a public library directly on
the libwabt target. This provides access to the necessary include files
from libwabt source files and automatically links jitbuilder into any
executable that depends on libwabt.