Implement constant-loading opcodes
The opcodes i32.const, i64.const, f32.const, and f64.const have now been
implemented. All that these opcodes do is push a single value onto the
stack for later use.

In the future, it's likely that the value stack would not be used and
these would be simplified, but for now we have decided to simply use the
value stack and leave such optimizations for later.

Closes #14