Add handling for value stack overflows
Prior to this change, JITted code was not making the proper checks to
ensure that the value stack was not exhausted before pushing values onto
it. This was resulting in a buffer overflow which could result in heap
corruption or a segfault under the right conditions.

JITted code now handles this correctly and returns
TrapValueStackExhausted if the value stack is exhausted when a push
operation occurs.

Fixes #23