Cache JITed function entry points
Instead of unconditionally JIT compiling a function when it is called, a table
mapping function offsets to JITed function entry points is first checked. If the
table already contains an entry point for the current function, then the
function is simply called. If no entry is found, the function will be JIT
compiled. If compilation fails, `nullptr` is set in the table.

Closes #21