Make JIT initialization/shutdown transparent
Previously, it was the responsibility of the program embedding the WABT
interpreter to ensure that the JIT was initialized and shutdown at the
correct times. However, this didn't really make sense since this is an
internal implementation detail of the JIT. For this reason, JIT
initialization and shutdown is now tied to the lifecycle of a
wabt::interp::Environment object.

Closes #10