Add extra template argument for op return type
Previously, the EmitBinaryOp and EmitUnaryOp functions only took a
single type and expected that the types of the arguments and the type of
the return value were all the same. This is a problem for comparison
opcodes, which always return an i32. Thus, an extra template argument
has been added to specify the return type.