Implement counted compilation
Previously, WASM functions were always JITted the first time they were
called. A more general way of deciding when to JIT a function is to
track the number of times it's called and compile when that value
reaches a certain threshold. It is now possible to do this by setting
Environment.jit_threshold to the desired number of calls after which to
JIT compile the function.

Closes #22