Add call stack information to ExecResult
For better debugging, call stacks are now returned in the ExecResult
retrieved from an Executor. This call stack will be printed if a
function traps in wasm-interp.

Additionally, a flag --no-stack-trace has been added to wasm-interp to
disable this functionality. This is useful for running the test suite
since the tests do not include stack traces in their expected output.