Fix JITted functions not saving the PC when trapping
Previously, the thread's PC would not be updated by the JIT when a trap
was encountered. Additionally, the calling interpreter would overwrite
the PC value saved, even if it were updated correctly. As a result of
this, the top frame of a stack trace would be incorrect in the presence
of a JITted function.