Dispatch the JIT on indirect_call from interpreter
Previously, due to an oversight, the JIT was not being dispatched when
an indirect_call opcode was being handled from the interpreter. This
meant that indirect calls from the interpreter would not cause JIT
compilation of functions and would always run the called function in the
interpreter, even if a JITted version of that function was already
present.

Fixes #89.