Use debug names in FunctionBuilder
When looking through OMR logs, it is often useful to be able to know
which function is being compiled. Previously, all WASM functions were
given a name of WASM_Function, which made it difficult to determine
which function was being referred to. Now that loading of function debug
names is implemented, it makes sense to switch to using them when
possible.