Fix host calls being unable to trap
Due to an oversight in the interpreter code in vanilla WABT, the host
call helper function is capable of returning trap error codes but these
codes are ignored by the interpreter. This same oversight was then
replicated in JITted code for consistency. However, our macrobenchmark
requires that trapping from host calls work correctly, as the exit_group
syscall must trap to stop execution.