Implement a virtual value stack for JIT-compiled functions
Previously, the value stack pushes and pops of each WASM opcode in a
function were actually performed on the physical value stack. Not only
did this require unnecessary memory reads and writes, it also required
that almost all opcodes perform bounds checking on the value stack. This
was causing problems with performance, both at runtime and when
compiling functions.

A better solution is to use a virtual value stack which emulates the
value stack at compile-time and turns stack pushes and pops into simple
value passing. This eliminates a large number of value stack bounds
checking and allows for more efficient passing of values within a
function.

Currently, there are still some uses of the physical value stack
remaining. For one, local variables are still stored on the value stack
where they would have been placed prior to this change for debugging
purposes. More importantly, function arguments and return values are
always placed on the physical value stack. It may be possible to avoid
doing this when calling one JIT-compiled function from another, but this
would require further work.

Closes #24