Update to the latest version of OMR
For various reasons, it makes sense for us to update to the latest
version of OMR for the JIT. Unfortunately, newer versions of OMR have a
different API for JitBuilder that doesn't include some of the
functionality that we need. For now, we're using a hacky workaround to
use the internal definitions of the JitBuilder classes rather than the
externally visible class definitions to allow access to those missing
functions.

The intention right now is for wasmjit-omr to eventually be updated to
use the JitBuilder API in its intended way once these problems are
corrected. However, it's easier for this to be done after updating OMR
rather than doing it at the same time.