Fix JIT physical stack indexing
Since the last time that upstream WABT changes were merged, the size of
elements on the physical stack has increased to accommodate v128 values
being placed on the physical value stack. While we don't intend to add
handling for actually using v128 values in JITted code at this time, we
do need to update code which indexes into the physical value stack to
take this into account when calculating the address to load from.