Fix JIT support for InterpDropKeep
Due to some changes to the interpreter in the name of multi-value
support, the InterpDropKeep opcode now uses a 32-bit integer for the
number of values to keep instead of an 8-bit integer. Because of this,
the JIT would start reading unaligned opcodes from the instruction
stream after encountering an InterpDropKeep. This has now been
corrected.