Display interpreter/JIT transitions in stack traces
Previously, the stack traces given when a trap occurred were quite bare
and only had information about what WebAssembly code caused the trap.
However, when debugging, it can be quite useful to know if the code in
question was running in the interpreter or in the JIT. To allow this,
stack traces will now note when the interpreter calls a JITted function
and when JITted code calls an interpreted function.

Closes #134