Remove branches from JITted code for f{32,64}.abs
Previously, the implementation of the f32.abs and f64.abs instructions
in JITted code required several branches to be added to the generated
code. This is obviously not ideal, since it should be a simple affair to
get the absolute value of a floating-point number. Instead, performing a
simple bitwise AND on the float allows the sign bit to be masked without
requiring any branches.