Improve format of stack traces
The format for stack traces has been improved to make testing somewhat
easier. Previously, each frame would always include the absolute PC at
which the trap occurred. However, this value depends on the order in
which the functions are put into the WASM file and is thus difficult to
use for testing.

Instead, the offset from the start of the function will now be used when
possible. While this still depends on how the interpreter internally
encodes the function, it no longer depends on the order of WASM
functions in the file.