Fix incorrect stack traces on JIT compilation failures
Previously, due to a stale PC, the stack trace printed on JIT
compilation failures from calls in JITted code was incorrect, as it
included an extra garbage stack frame compared to what would be seen if
the failure occurred in the interpreter.

This has been fixed by normalizing the stack traces for JIT compilation
failure to include a topmost frame showing the function being compiled.
This should clearly indicate to the user which function is failing to
JIT compile. For further clarity, this fake frame will be printed with a
clarification noting that the failure occurred during JIT compilation.

Fixes #135