Add temporary fix for stack merge failure
Due to a change in upstream OMR a while back, virtual stack merging
stopped working correctly, causing strange results when returning
results from if blocks. This was caused by JitBuilder not properly
implementing a fallback for calling VirtualMachineState::MergeInto if no
client callback is registered, as is the case for us since we aren't yet
able to use the JitBuilder client API due to compatibility problems.

A very hacky workaround has been added which registers a simple client
callback on all WabtState objects that will forward the call to the
actual implementation of WabtState::MergeInto. This workaround is
intended to be removed when either the upstream JitBuilder problem is
resolved or we're eventually able to move to the JitBuilder client API.

Fixes #151