Add tests for multi-value functionality
Previously, there weren't any tests for support for multi-return from
functions or blocks. This was a result of these being implicitly
supported due to the way we were handling the virtual stack. However, in
order to ensure that this functionality continues to work in the future,
new tests have been added to make use of this functionality.

Closes #133