Fix Thread::in_jit_ not being cleared between calls
After encountering a trap in JITted code, the interpreter will have to
return control to the caller while Thread::in_jit_ is set to true.
However, when starting to run a new interpreted function, this flag is
never cleared, leading a trap in purely interpreted code to look like it
happened in JITted code if no further JITted code is run before the next
function traps.