Fix JitBuilder include path
Previously, the CMakeLists.txt file for the JIT was setting the
JitBuilder include directory relative to ${CMAKE_SOURCE_DIR}. While this
was correct when building wasmjit-omr on its own, it is not correct if
wasmjit-omr is being included in another CMake project through the use
of add_subdirectory. Instead, using ${CMAKE_CURRENT_SOURCE_DIR} to set
the JitBuilder include path will do the right thing for all use cases.