Use function indices instead of offsets for call opcode
Previously, the WABT interpreter would record a call opcode with an
immediate representing the offset of the instruction to jump to. While
this is generally good for the interpreter, our JIT needs to keep
certain function metadata available when making calls. Having only the
offset of the function's first instruction available meant that we
needed to keep a table mapping offsets to functions.

With this change, it should now be possible to eliminate this and simply
keep metadata alongside WABT's other function metadata, though this is
not yet done.