Improve function dispatch from JITted code
Previously, in order to call another function from within a JITted
function, it was necessary to call a helper function. This incurred an
unnecessary amount of overhead when calling JITted functions from other
JITted functions.

In order to make calls more lightweight, a new dispatch system has been
introduced. Instead of calling a helper, direct function calls from
JITted code will now instead be made via a table accessible to JITted
code. For interpreted and host functions, this table will be populated
with function pointers to thunks which will perform the necessary
transition. Once a function has been JITted, its entry in the table will
be replaced with a pointer directly to the JITted function, allowing the
helper to be bypassed entirely.