Add test cases for more complex call sequences
With the addition of more complicated dispatch for both interpreter to
JIT and JIT to interpreter transitions, several bugs have been
discovered that didn't cause any test failures. This was due to the lack
of more complex call chains and sequences in any of the tests.

To alleviate this, a new test has been added which tests several more
complex call sequences to ensure that they work as expected.