Remove use of std::vector in Memory
Previously, the interpreter data structure representing WebAssembly
linear memory was using a std::vector<char> as a smart pointer to point
to its data. This created a problem for the JIT, since std::vector<char>
is not standard layout and thus cannot be accessed from JITted code. As
a result of this, the JIT had to call a helper every time it wanted to
perform a memory operation, which was seriously degrading performance.

To alleviate this, a new MemoryData structure has been added that is
used in place of std::vector<char>. For the most part, this structure
mimics the basic vector APIs, allowing it to be replaced seamlessly.
Importantly, this structure is standard layout, meaning that the JIT can
now read its fields without using a helper.