Fix incorrect aliasing on linear memory accesses
Previously, linear memory accesses from JITted code were done using the
LoadAt and StoreAt services. However, recent improvements in the JIT
have allowed more flexibility for the optimizer and have thus exposed a
latent bug with this approach: JitBuilder creates indirect loads/stores
assuming that memory accesses for LoadAt/StoreAt calls with different
types cannot alias each other. It's unclear whether the bug lies with
JitBuilder or with our JIT, as this is admittedly not something that
typically occurs. Nevertheless, a workaround has been added to the JIT
so that all accesses to linear memory now use LoadIndirect and
StoreIndirect with a union to ensure correct aliasing.